+++ api/helm-app/HelmAppRestHandler.go
@@ -292,49 +292,14 @@ func (handler *HelmAppRestHandlerImpl) UpdateApplication(w http.ResponseWriter, 
 	}
 	//RBAC enforcer Ends
 
-	installedApp, err := handler.appStoreDeploymentCommonService.GetInstalledAppByClusterNamespaceAndName(appIdentifier.ClusterId, appIdentifier.Namespace, appIdentifier.ReleaseName)
+	// update application externally
+	res, err := handler.helmAppService.UpdateApplication(context.Background(), appIdentifier, request)
 	if err != nil {
 		common.WriteJsonResp(w, err, nil, http.StatusInternalServerError)
 		return
 	}
-
-	updateReleaseRequest := &InstallReleaseRequest{
-		ValuesYaml: request.GetValuesYaml(),
-		ReleaseIdentifier: &ReleaseIdentifier{
-			ReleaseNamespace: appIdentifier.Namespace,
-			ReleaseName:      appIdentifier.ReleaseName,
-		},
-	}
-
-	var res *openapi.UpdateReleaseResponse
-
-	if installedApp != nil {
-		chartInfo := installedApp.InstallAppVersionChartDTO
-		chartRepoInfo := chartInfo.InstallAppVersionChartRepoDTO
-		updateReleaseRequest.ChartName = chartInfo.ChartName
-		updateReleaseRequest.ChartVersion = chartInfo.ChartVersion
-		updateReleaseRequest.ChartRepository = &ChartRepository{
-			Name:     chartRepoInfo.RepoName,
-			Url:      chartRepoInfo.RepoUrl,
-			Username: chartRepoInfo.UserName,
-			Password: chartRepoInfo.Password,
-		}
-		res, err = handler.helmAppService.UpdateApplicationWithChartInfo(context.Background(), appIdentifier.ClusterId, updateReleaseRequest)
-		if err != nil {
-			common.WriteJsonResp(w, err, nil, http.StatusInternalServerError)
-			return
-		}
-	} else {
-		res, err = handler.helmAppService.UpdateApplication(context.Background(), appIdentifier, request)
-		if err != nil {
-			common.WriteJsonResp(w, err, nil, http.StatusInternalServerError)
-			return
-		}
-	}
-
 	common.WriteJsonResp(w, err, res, http.StatusOK)
 }
-
 
 func (handler *HelmAppRestHandlerImpl) checkHelmAuth(token string, object string) bool {
 	if ok := handler.enforcer.Enforce(token, casbin.ResourceHelmApp, casbin.ActionGet, strings.ToLower(object)); !ok {
